Admin Auth Config options & cleanup
