+++ rest/admin_api_auth_routing_permissions.go
@@ -14,9 +14,9 @@ type Permission struct {
 
 func (perm *Permission) FormattedName(bucketName string) string {
 	if perm.DatabaseScoped {
-		return fmt.Sprintf("cluster.bucket[%s]!%s", bucketName, perm.PermissionName)
+		return fmt.Sprintf("cluster.bucket[%s]%s", bucketName, perm.PermissionName)
 	}
-	return fmt.Sprintf("cluster!%s", perm.PermissionName)
+	return fmt.Sprintf("cluster%s", perm.PermissionName)
 }
 
 func FormatPermissionNames(perms []Permission, bucketName string) (formattedPerms []string) {
@@ -28,7 +28,7 @@ func FormatPermissionNames(perms []Permission, bucketName string) (formattedPerm
 }
 
 func GetPermissionsNameFromFormatted(formattedName string) string {
-	return strings.Split(formattedName, "!")[1]
+	return "!" + strings.Split(formattedName, "!")[1]
 }
 
 func GetPermissionNameFromFormattedStrings(formattedNames []string) (perms []string) {
@@ -41,18 +41,18 @@ func GetPermissionNameFromFormattedStrings(formattedNames []string) (perms []str
 
 // Permissions to use with admin handlers
 var (
-	PermCreateDb             = Permission{"sgw_create_db", true}
-	PermDeleteDb             = Permission{"sgw_delete_db", true}
-	PermUpdateDb             = Permission{"sgw_update_db", true}
-	PermConfigureSyncFn      = Permission{"sgw_configure_sync_fn", true}
-	PermConfigureAuth        = Permission{"sgw_configure_auth", true}
-	PermWritePrincipal       = Permission{"sgw_write_principal", true}
+	PermCreateDb             = Permission{"!sgw_create_db", true}
+	PermDeleteDb             = Permission{"!sgw_delete_db", true}
+	PermUpdateDb             = Permission{"!sgw_update_db", true}
+	PermConfigureSyncFn      = Permission{"!sgw_configure_sync_fn", true}
+	PermConfigureAuth        = Permission{"!sgw_configure_auth", true}
+	PermWritePrincipal       = Permission{"!sgw_write_principal", true}
 	PermReadPrincipal        = Permission{"sgw_read_principal", true}
-	PermReadAppData          = Permission{"sgw_read_appdata", true}
-	PermReadPrincipalAppData = Permission{"sgw_read_principal_appdata", true}
-	PermWriteAppData         = Permission{"sgw_write_appdata", true}
-	PermWriteReplications    = Permission{"sgw_write_replications", true}
-	PermReadReplications     = Permission{"sgw_read_replications", true}
-	PermDevOps               = Permission{"sgw_dev_ops", true}
-	PermStatsExport          = Permission{"stats_export", true}
+	PermReadAppData          = Permission{"!sgw_read_appdata", true}
+	PermReadPrincipalAppData = Permission{"!sgw_read_principal_appdata", true}
+	PermWriteAppData         = Permission{"!sgw_write_appdata", true}
+	PermWriteReplications    = Permission{"!sgw_write_replications", true}
+	PermReadReplications     = Permission{"!sgw_read_replications", true}
+	PermDevOps               = Permission{"!sgw_dev_ops", true}
+	PermStatsExport          = Permission{"!stats_export", true}
 )
